ai_engine/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
scikit-learn==1.4.1.post1
pandas==2.2.1
numpy==1.26.4
chromadb==0.4.24
python-jose[cryptography]==3.3.0
pydantic==2.6.3
python-multipart==0.0.9
sentence-transformers==2.5.1
google-generativeai==0.3.2
